requirements.txt
pyyaml==3.12
pandas==0.18.1
matplotlib==1.5.1
seaborn==0.7.1
slacker==0.9.25
